The Jupyter Lab implementation of the ICommandPalette interface ought to be private.
