+++ src/commandpalette/plugin.ts
@@ -3,6 +3,11 @@
 | Distributed under the terms of the Modified BSD License.
 |----------------------------------------------------------------------------*/
 
+
+import {
+  DisposableDelegate, IDisposable
+} from 'phosphor/lib/core/disposable';
+
 import {
   CommandPalette
 } from 'phosphor/lib/ui/commandpalette';
@@ -12,8 +17,54 @@ import {
 } from '../application';
 
 import {
-  Palette, ICommandPalette
+  ICommandPalette, IPaletteItem
 } from './';
+
+
+/**
+ * A thin wrapper around the `CommandPalette` class to conform with the
+ * JupyterLab interface for the application-wide command palette.
+ */
+class Palette {
+  /**
+   * Create a palette instance.
+   */
+  constructor(palette: CommandPalette) {
+    this._palette = palette;
+  }
+
+  /**
+   * The placeholder text of the command palette's search input.
+   */
+  set placeholder(placeholder: string) {
+    this._palette.inputNode.placeholder = placeholder;
+  }
+  get placeholder(): string {
+    return this._palette.inputNode.placeholder;
+  }
+
+  /**
+   * Add a command item to the command palette.
+   *
+   * @param options - The options for creating the command item(s).
+   *
+   * @returns A disposable that will remove the item from the palette.
+   */
+  addItem(options: IPaletteItem): IDisposable {
+    let item = this._palette.addItem(options as CommandPalette.IItemOptions);
+    return new DisposableDelegate(() => this._palette.removeItem(item));
+  }
+
+  /**
+   * Focus the search input node of the command palette.
+   */
+  focus(): void {
+    this._palette.inputNode.focus();
+    this._palette.inputNode.select();
+  }
+
+  private _palette: CommandPalette = null;
+}
 
 
 /**
